+++ CHANGES.rst
@@ -55,6 +55,12 @@ Changelog
   decided on a meeting, instead just refreshing the faceted, the faceted is
   refreshed and the screen scrolls to the modified item.
   [gbastien]
+- Completed `MeetingConfig.validate_itemWFValidationLevels` to check that the
+  `itemcreated` state always exists as first element (even if may be disabled),
+  check also that every `back_transition` back transition identifier starts with
+  `back` and that format of identifier columns (`state`, `leading_transition`,
+  `back_transition` must be only alphanumeric) is correct.
+  [gbastien]
 
 4.2b13 (2021-07-16)
 -------------------
@@ -524,9 +530,9 @@ Changelog
   if not redefined on item, and when redefined, it takes precedence over what
   is defined in meeting.
   [gbastien]
-- Completed `MeetingConfig.itemWFValidationLevels` to check, when an state is
-  removed, if it is not used by a workflowAdaptation.
-  For example workflowAdaptation `waiting_advices` may creates state
+- Completed `MeetingConfig.validate_itemWFValidationLevels` to check, when a
+  state is removed, if it is not used by a workflowAdaptation.
+  For example workflowAdaptation `waiting_advices` may create state
   `proposed_waiting_advices`, in this case state `proposed` can not be removed
   if some items still in `proposed_waiting_advices`.
   We check every states id beginning with removed states or containing